backend/requirements.txt
fastapi
uvicorn
sqlalchemy
cx_Oracle
python-multipart
aiofiles
zipstream-new